requirements.txt
ccxt
numpy
python-telegram-bot==20.7
